Group Invoice Create Edit

+++ app/Http/Controllers/Purchase/GroupInvoiceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Purchase;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\StoreGroupInvoicePurchaseOrder;
 use App\Models\PurchaseItem;
 use App\Models\PurchaseOrder;
 use App\TemporaryPurchaseGroupItem;
@@ -40,35 +41,38 @@ class GroupInvoiceController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreGroupInvoicePurchaseOrder $request)
     {
-        $purchase_order = new PurchaseOrder();
-        $purchase_order->supplier_id = 0;
-        $purchase_order->purchase_no = $request->purchase_no;
-        $purchase_order->purchase_date = $request->purchase_date;
-        $purchase_order->purchase_representative_id = $request->purchase_representative_id;
-        $purchase_order->user_id = auth()->user()->id ?? 0;
-        $purchase_order->total_amount = $request->total_amount;
-        $purchase_order->order_status = $request->order_status;
-        $purchase_order->invoice_status = 'group_invoice';
-        $purchase_order->save();
-        $purchase_order_id = $purchase_order->id;
+        if ($request->productFields) {
+            $purchase_order = new PurchaseOrder();
+            $purchase_order->supplier_id = 0;
+            $purchase_order->purchase_no = $request->purchase_no;
+            $purchase_order->purchase_date = $request->purchase_date;
+            $purchase_order->purchase_representative_id = $request->purchase_representative_id;
+            $purchase_order->user_id = auth()->user()->id ?? 0;
+            $purchase_order->total_amount = $request->total_amount;
+            $purchase_order->order_status = $request->order_status;
+            $purchase_order->invoice_status = 'group_invoice';
+            $purchase_order->save();
+            $purchase_order_id = $purchase_order->id;
 
-        foreach ($request->productFields as $key => $value) {
-            $insert[$key]['brand_id'] = $value['brand_id'];
-            $insert[$key]['type_of_model_id'] = $value['type_of_model_id'];
-            $insert[$key]['qty'] = $value['qty'];
-            $insert[$key]['cif_usd'] = $value['cif_usd'];
-            $insert[$key]['description'] = $value['description'];
-            $insert[$key]['purchase_order_id'] = $purchase_order_id;
-            $insert[$key]['purchase_order_original_id'] = $value['purchase_order_id'];
-            $insert[$key]['purchase_item_id'] = $value['purchase_item_id'];
-            $insert[$key]['created_at'] =  date('Y-m-d H:i:s');
-            $insert[$key]['updated_at'] =  date('Y-m-d H:i:s');
+            foreach ($request->productFields as $key => $value) {
+                $insert[$key]['brand_id'] = $value['brand_id'];
+                $insert[$key]['type_of_model_id'] = $value['type_of_model_id'];
+                $insert[$key]['qty'] = $value['qty'];
+                $insert[$key]['cif_usd'] = $value['cif_usd'];
+                $insert[$key]['description'] = $value['description'];
+                $insert[$key]['purchase_order_id'] = $purchase_order_id;
+                $insert[$key]['purchase_order_original_id'] = $value['purchase_order_id'];
+                $insert[$key]['purchase_item_id'] = $value['purchase_item_id'];
+                $insert[$key]['created_at'] =  date('Y-m-d H:i:s');
+                $insert[$key]['updated_at'] =  date('Y-m-d H:i:s');
+            }
+            PurchaseItem::insert($insert);
+            TemporaryPurchaseGroupItem::where('session_id', session()->getId())->delete();
+            return redirect()->back()->with('success', 'Your processing has been completed.');
         }
-        PurchaseItem::insert($insert);
-        TemporaryPurchaseGroupItem::where('session_id', session()->getId())->delete();
-        return redirect()->back()->with('success', 'Your processing has been completed.');
+        return redirect()->back()->with('error', 'Error.');
     }
 
     /**
@@ -90,9 +94,6 @@ class GroupInvoiceController extends Controller
      */
     public function edit($id)
     {
-        // $suppliers = Supplier::all();
-        // $brands = Brand::all();
-
         $sales_persons = User::all();
         $purchase_order_info = PurchaseOrder::findOrFail($id);
         $purchase_orders = PurchaseOrder::all();
@@ -109,7 +110,35 @@ class GroupInvoiceController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        $purchase_order = PurchaseOrder::findOrFail($id);
+        $purchase_order->supplier_id = 0;
+        $purchase_order->purchase_no = $request->purchase_no;
+        $purchase_order->purchase_date = $request->purchase_date;
+        $purchase_order->purchase_representative_id = $request->purchase_representative_id;
+        $purchase_order->user_id = auth()->user()->id ?? 0;
+        $purchase_order->total_amount = $request->total_amount;
+        $purchase_order->order_status = $request->order_status;
+        $purchase_order->invoice_status = 'group_invoice';
+        $purchase_order->update();
+        $purchase_order_id = $purchase_order->id;
+
+        if ($request->productFields) {
+            foreach ($request->productFields as $key => $value) {
+                $insert[$key]['brand_id'] = $value['brand_id'];
+                $insert[$key]['type_of_model_id'] = $value['type_of_model_id'];
+                $insert[$key]['qty'] = $value['qty'];
+                $insert[$key]['cif_usd'] = $value['cif_usd'];
+                $insert[$key]['description'] = $value['description'];
+                $insert[$key]['purchase_order_id'] = $purchase_order_id;
+                $insert[$key]['purchase_order_original_id'] = $value['purchase_order_id'];
+                $insert[$key]['purchase_item_id'] = $value['purchase_item_id'];
+                $insert[$key]['created_at'] =  date('Y-m-d H:i:s');
+                $insert[$key]['updated_at'] =  date('Y-m-d H:i:s');
+            }
+            PurchaseItem::insert($insert);
+            TemporaryPurchaseGroupItem::where('session_id', session()->getId())->delete();
+        }
+        return redirect()->back()->with('success', 'Your processing has been completed.');
     }
 
     /**
